+++ R/SPR.R
@@ -3,19 +3,24 @@
 #' @export
 SPRWarning <- function (parent, child, error) {
   warning ("No SPR operation performed.\n  > ", error)
-  return(list(parent, child))
+  
+  # Return:
+  list(parent, child)
 }
 
-#' Subtree Pruning and Rearrangement 
+#' Subtree Pruning and Rearrangement (SPR)
 #'
 #' Perform one \acronym{SPR} rearrangement on a tree
 #' 
-#' Equivalent to phangorn's kSPR, but faster.
+#' Equivalent to phangorn's `kSPR`, but faster.
 #' Note that rearrangements that only change the position of the root WILL be returned by 
 #' \code{SPR}.  If the position of the root is irrelevant (as in Fitch parsimony, for example)
 #' then this function will occasionally return a functionally equivalent topology.  
 #' \code{RootIrrelevantSPR} will search tree space more efficiently in these cases.
 #' Branch lengths are not (yet) supported.
+#'
+#' All nodes in a tree must be bifurcating; [ape::collapse.singles] and
+#' [ape::multi2di] may help.
 #'
 #' @template treeParam
 #' @param edgeToBreak the index of an edge to bisect, generated randomly if not specified.
@@ -40,20 +45,23 @@ SPRWarning <- function (parent, child, error) {
 SPR <- function(tree, edgeToBreak = NULL, mergeEdge = NULL) {
   if (is.null(treeOrder <- attr(tree, 'order')) || treeOrder != 'preorder') tree <- Preorder(tree)
   edge <- tree$edge
+  parent <- edge[, 1]
+  StopUnlessBifurcating(parent)
   if (!is.null(edgeToBreak) && edgeToBreak == -1) {
-    parent <- edge[, 1]
     child <- edge[, 2]
     nEdge <- length(parent)
     stop('Negative edgeToBreak not yet supported; on TODO list for next release')
     notDuplicateRoot <- NonDuplicateRoot(parent, child, nEdge)
-    return(unique(unlist(lapply(which(notDuplicateRoot), AllSPR,
+    # Return:
+    unique(unlist(lapply(which(notDuplicateRoot), AllSPR,
       parent=parent, child=child, nEdge=nEdge, notDuplicateRoot=notDuplicateRoot),
-      recursive=FALSE))) # TODO the fact that we need to use `unique` indicates that 
+      recursive=FALSE)) # TODO the fact that we need to use `unique` indicates that 
                          #      we're being inefficient here.
   } else {
-    tree$edge <- ListToMatrix(SPRSwap(edge[, 1], edge[, 2], edgeToBreak=edgeToBreak, 
+    tree$edge <- ListToMatrix(SPRSwap(parent, edge[, 2], edgeToBreak=edgeToBreak, 
                                       mergeEdge=mergeEdge))
-    return(tree)
+    # Return:
+    tree
   }
 }
 
@@ -176,7 +184,8 @@ AllSPR <- function (parent, child, nEdge, notDuplicateRoot, edgeToBreak) {
       newChild [brokenEdgeSister] <- child[mergeEdge]
       newParent[brokenEdge | brokenEdgeSister] <- child[brokenEdgeSister]
       newChild[mergeEdge] <- child[brokenEdgeSister]
-      return(RenumberTree(newParent, newChild, nEdge))
+      # Return:
+      RenumberTree(newParent, newChild, nEdge)
     }) # lapply faster than vapply
   } else {
     newEdges <- lapply(mergeEdges, function (mergeEdge) {
@@ -184,10 +193,12 @@ AllSPR <- function (parent, child, nEdge, notDuplicateRoot, edgeToBreak) {
       newParent[brokenEdgeSister] <- parent[brokenEdgeParent]
       newParent[brokenEdgeParent] <- newParent[mergeEdge]
       newParent[mergeEdge] <- brokenEdge.parentNode
-      return(RenumberTree(newParent, child, nEdge))
+      # Return:
+      RenumberTree(newParent, child, nEdge)
     }) # lapply faster than vapply
   }
-  return(lapply(newEdges, function (newEdge) {tree$edge <- newEdge; tree}))
+  # Return:
+  lapply(newEdges, function (newEdge) {tree$edge <- newEdge; tree})
 }
 
 #' Rooted SPR 
@@ -290,5 +301,7 @@ RootedSPRSwap <- function (parent, child, nEdge = length(parent), nNode = nEdge 
   
   #####Assert(identical(unique(table(parent)), 2L))
   #####Assert(identical(unique(table(child)),  1L))
-  return(RenumberEdges(parent, child, nEdge))  
+  
+  # Return:
+  RenumberEdges(parent, child, nEdge)
 }